+++ api/.env.example
@@ -1,3 +1,4 @@
+# Used by docker compose when starting postgres
 POSTGRES_USER=postgres
 POSTGRES_PASSWORD=postgres
 POSTGRES_DB=rc4laundry_test
