+++ api/.gitignore
@@ -1,1 +1,1 @@
-.env.local
+.env
