+++ api/routers/__init__.py
@@ -1,0 +1,1 @@
+from .machine import router as machine
